[ci] reorder test_resource_limits test
