+++ python/ray/data/BUILD
@@ -483,6 +483,14 @@ py_test(
 )
 
 py_test(
+    name = "test_streaming_executor_resource_limits",
+    size = "medium",
+    srcs = ["tests/test_streaming_executor_resource_limits.py"],
+    tags = ["team:data", "exclusive"],
+    deps = ["//:ray_lib", ":conftest"],
+)
+
+py_test(
     name = "test_streaming_integration",
     size = "medium",
     srcs = ["tests/test_streaming_integration.py"],
